Time for bed.

+++ app/database/migrations/2013_09_02_172701_create_users_table.php
@@ -17,7 +17,7 @@ class CreateUsersTable extends Migration {
 			$table->string('email')->unique();
 			$table->string('password');
 			$table->boolean('is_guest');
-			$table->integer('ip_address'); // see http://stackoverflow.com/questions/6427786/ip-address-storing-in-mysql-database
+			$table->bigInteger('ip_address'); // see http://stackoverflow.com/questions/6427786/ip-address-storing-in-mysql-database
 			//$table->timestamp('created_at')->default(date("Y-m-d H:i:s"));
 			//$table->timestamp('modified_at')->default('0000-00-00 00:00');
 			$table->timestamps();
